// src/app/app/cars/interfaces/car.ts
export interface Car {
    
    id: number;
    amount: number;
    carStateNumberPlate: string;
    mileage: string;
    status: string;
    year: string;
    brand: string;
    model: string;
    bodyType: string;
    color: string;
    city: string;
    imagePaths: string[]; 
    branchId: number;
    rentalId: number;

}